modify the original one to reflect my .cpp lib

+++ content/tanya-lib/template.h
@@ -1,3 +1,7 @@
+/**
+ * Author: Tanya
+ * Description: TODO
+ */
 #include<bits/stdc++.h>
 using namespace std;
 void __print(int x) {cerr << x;}
@@ -53,7 +57,7 @@ template<class T> using pqg = priority_queue<T, vector<T>, greater<T>>;
 #define FORd(i,a,b) for (int i = (b)-1; i >= a; i--)
 #define F0Rd(i,a) for (int i = (a)-1; i >= 0; i--)
 #define trav(a,x) for (auto& a : x)
-#define uid(a, b) uniform_int_distribution<int>(a, b)(rng)
+#define uid(a, b) uniform_int_distribution<long long>(a, b)(rng)
  
 #define sz(x) (int)(x).size()
 #define mp make_pair
@@ -71,7 +75,7 @@ template<class T> bool ckmax(T& a, const T& b) { return a < b ? a = b, 1 : 0; }
 mt19937 rng(chrono::steady_clock::now().time_since_epoch().count());
 
 const char nl = '\n';
-const int N =2e5+1;
+const int N = 2e5+1;
 const int INF = 1e9+7;
 const long long LINF = 1e18+7;
